refactored most redis crud into peer_crud.go
