+++ storage/redistore/peer_crud.go
@@ -1,11 +1,22 @@
 package redistore
 
 import (
+	"encoding/binary"
 	"fmt"
+	"net"
 	"time"
 
 	"github.com/chihaya/chihaya/bittorrent"
+	"github.com/garyburd/redigo/redis"
 )
+
+func decodePeerKey(pk string) bittorrent.Peer {
+	return bittorrent.Peer{
+		ID:   bittorrent.PeerIDFromString(string(pk[:20])),
+		Port: binary.BigEndian.Uint16([]byte(pk[20:22])),
+		IP:   net.IP(pk[22:]),
+	}
+}
 
 func addAndCleanPeer(s *peerStore, infoHash bittorrent.InfoHash, peerType string, pk serializedPeer) error {
 	nameSpacedKey := addNameSpace(fmt.Sprintf("%s%s", peerType+":", infoHash))
@@ -29,4 +40,23 @@ func remPeers(s *peerStore, infoHash bittorrent.InfoHash, peerType string, pk se
 		return err
 	}
 	return nil
+}
+
+func getPeers(s *peerStore, infoHash bittorrent.InfoHash, peerType string, numWant int, peers []bittorrent.Peer, announcer bittorrent.Peer) ([]bittorrent.Peer, error) {
+	peerList, err := redis.Strings(s.conn.Do("ZRANGE",
+		addNameSpace(fmt.Sprintf("%s%s", peerType+":", infoHash)), 0, -1))
+	if err != nil {
+		return nil, err
+	}
+	for _, p := range peerList {
+		if numWant == len(peers) {
+			break
+		}
+		decodedPeer := decodePeerKey(p)
+		if decodedPeer.Equal(announcer) {
+			continue
+		}
+		peers = append(peers, decodedPeer)
+	}
+	return peers, nil
 }
